refactor: update select, dropdown, toggle components to new form pattern
