+++ components/ui/toggle/material.tsx
@@ -1,14 +1,21 @@
 'use client';
 
-// import { COMPONENT_IDS } from '@/core/system/componentRegistry';
+import FormatAlignCenterIcon from '@mui/icons-material/FormatAlignCenter';
+import FormatAlignJustifyIcon from '@mui/icons-material/FormatAlignJustify';
+import FormatAlignLeftIcon from '@mui/icons-material/FormatAlignLeft';
+import FormatAlignRightIcon from '@mui/icons-material/FormatAlignRight';
+import FormatBoldIcon from '@mui/icons-material/FormatBold';
+import FormatColorFillIcon from '@mui/icons-material/FormatColorFill';
+import FormatItalicIcon from '@mui/icons-material/FormatItalic';
+import FormatUnderlinedIcon from '@mui/icons-material/FormatUnderlined';
 import { ToggleButton, ToggleButtonGroup } from '@mui/material';
 import * as React from 'react';
 
+import { FormFieldConfig } from '@/components/shared/form/FormField';
 // import ToggleProps from '@mui/material/ToggleButton'; // Q: Should I be using this instead of the custom ButtonProps?
 import { MaterialComponentColor, MaterialComponentSize } from '@/core/constants/material';
 
-import { ToggleProps } from '..';
-import { exclusiveSampleData, nonExclusiveSampleData } from './config';
+import { ToggleProps } from './componentRegistry';
 
 /*
  *  ---- MATERIAL UI BUTTONS ----
@@ -48,6 +55,89 @@ export interface MaterialToggleProps extends ToggleProps {
   disableFocusRipple?: boolean;
   fullWidth?: boolean;
 }
+
+/*
+ * ---- FORM CONFIGURATION ----
+ */
+export const materialToggleFields: FormFieldConfig[] = [
+  {
+    type: 'dropdown',
+    label: 'Size',
+    prop: 'size',
+    options: Object.values(MaterialComponentSize),
+    default: MaterialComponentSize.Medium
+  },
+  {
+    type: 'dropdown',
+    label: 'Color',
+    prop: 'color',
+    options: Object.values(MaterialComponentColor),
+    default: MaterialComponentColor.Primary
+  },
+  {
+    type: 'dropdown',
+    label: 'Variant',
+    prop: 'variant',
+    options: Object.values(MaterialToggleOrientation),
+    default: MaterialToggleOrientation.Horizontal
+  },
+  {
+    type: 'toggle',
+    label: 'Exclusive',
+    prop: 'isExclusive',
+    default: false
+  },
+  {
+    type: 'toggle',
+    label: 'Checked',
+    prop: 'checked',
+    default: false
+  },
+  {
+    type: 'toggle',
+    label: 'Full Width',
+    prop: 'fullWidth',
+    default: false
+  },
+  {
+    type: 'toggle',
+    label: 'Disabled',
+    prop: 'disabled',
+    default: false
+  },
+  {
+    type: 'toggle',
+    label: 'Disabled Ripple',
+    prop: 'disableRipple',
+    default: false
+  },
+  {
+    type: 'toggle',
+    label: 'Disable Focus Ripple',
+    prop: 'disableFocusRipple',
+    default: false
+  }
+];
+
+/*
+ * ---- SAMPLE DATA SETS ----
+ */
+
+// For rendering ToggleButtonGroup with nonexclusive selections
+export const nonExclusiveSampleData = [
+  { value: 'bold', label: 'Bold', icon: <FormatBoldIcon /> },
+  { value: 'italic', label: 'Italic', icon: <FormatItalicIcon /> },
+  { value: 'underlined', label: 'Underlined', icon: <FormatUnderlinedIcon /> },
+  { value: 'color', label: 'Color Fill', icon: <FormatColorFillIcon />, disabled: true }
+];
+
+// For rendering ToggleButtonGroup with exclusive selections
+export const exclusiveSampleData = [
+  { value: 'left', label: 'Left', icon: <FormatAlignLeftIcon /> },
+  { value: 'center', label: 'Center', icon: <FormatAlignCenterIcon /> },
+  { value: 'right', label: 'Right', icon: <FormatAlignRightIcon /> },
+  { value: 'justify', label: 'Justify', icon: <FormatAlignJustifyIcon />, disabled: true }
+];
 
 /*
  * ---- DEFAULT COMPONENT EXPORT ----
